css changes to font/spacing of roster
changes a few font weights and sizes. increased sizing and padding on a
few elements.

changes needed to other docs:
roomNameSpan -> remove open-sans-bold class

roomNameSpan -> dont know how you want to do it, but the text should
only be 1 line tall, with overflow hidden, and ... appended if name is
hidden || we can change this if its a pain in the ass

bottom border isnt showing up on .panel-heading when tasks are shown

if directed at team, or no direction given on text, remove  @Team PSF -
only if specific player is specified should this show... seeing that
light blue i have a strong feeling that nobody is going to like that
color, lol